Tutor_version_0.1/pyproject.toml
[project]
name = "tutor-version-0-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.8.0",
    "openai-agents",
]

[tool.uv.sources]
openai-agents = { git = "https://github.com/openai/openai-agents-python", rev = "main" }

[tool.uv.workspace]
members = ["MCP_tools"]
